+++ cmake/modules/Bridge.cmake
@@ -9,6 +9,7 @@ target_sources(kwik_bridge
             modules/bridge/prop_bus.cppm
             modules/bridge/binding_registry.cppm
             modules/bridge/bindings.cppm
+            modules/bridge/theme_bridge.cppm
     
     PRIVATE
         src/bridge/props_parser.cpp
@@ -16,6 +17,7 @@ target_sources(kwik_bridge
         src/bridge/prop_bus.cpp
         src/bridge/binding_registry.cpp
         src/bridge/bindings.cpp
+        src/bridge/theme_bridge.cpp
 )
 target_link_libraries(kwik_bridge
     PRIVATE